modules/serwis ekonomiczny/serwis_ekonomiczny_IO/requirements.txt
Flask>=2.3.0
Flask-SQLAlchemy>=3.0.0
python-dotenv>=1.0.0
pymysql>=1.1.2
requests>=2.30.0
pandas>=2.0.0
numpy>=1.27.0
matplotlib>=3.8.0